requirements.txt
streamlit
groq
pdfplumber
pydantic
pillow
paddleocr
